pyproject.toml
[project]
name = "noslacking"
version = "0.1.0"
description = "CLI tool to migrate a Slack workspace to Google Chat using APIs — no exports needed"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
keywords = ["slack", "google-chat", "migration", "workspace"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "typer>=0.14.0",
    "rich>=13.9.0",
    "slack-sdk>=3.33.0,<4.0.0",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.36",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "arrow>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[project.scripts]
noslacking = "noslacking.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/noslacking"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM"]
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-v", "--tb=short"]
